+++ app/containers/FolderActionBar.ts
@@ -3,6 +3,7 @@ import FolderActionBar from '../components/FolderActionBar';
 import { prepareDelete, startRename, startCreate } from '../actions/currentNode';
 import { createInCurrent } from '../actions/edit';
 import { open as openImport } from '../actions/fileImport';
+import { open as openExport } from '../actions/fileExport';
 import { openCurrent as openPermissions } from '../actions/authorizedUsers';
 import { browseForAdd } from '../actions/external';
 import {ROOT_ID} from '../domain/Node';
@@ -23,5 +24,6 @@ export default connect((state: RootState) => ({
   onEditPermissions: () => dispatch(openPermissions()),
   onAddExternal: () => dispatch(browseForAdd()),
   onImport: () => dispatch(openImport()),
+  onExport: () => dispatch(openExport()),
   onCopyLink: () => dispatch((_, getState) => copyStashLink(getState().currentNode.nodeId!))
 }))(FolderActionBar);