+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Stash links can now also be shared for folders
 - A shiny new splashscreen now tries to mask the unfathomably long loading times
 - KeePass import now creates entries for binary attachments
+- **KeePass export** is now available in the folder bars menu
 
 #### Changes
 - The *My Favorites* link and the tool icons now remain visible when scrolling